crontab: Do a pg_dump of selected tables from the database every day
It's useful for getting a developer up-and-running with Pombola to
have a database dump available; it would also be useful as part of
the provisioning of a Vagrant box to be able to restore a database
dump to get a realistic site.  This commit gets cron to do a pg_dump
of the database daily which can be used in both these situations.

The risk here is that there is personal information in the database
(e.g. the auth_user table, the django_session table, feedback
submitted to the site, etc.)  To avoid this we specify only the
tables we want to dump, and have looked through the data in these
by hand.  There's still room for error here, but it's safer than
trying to exclude every table that might include personal data.

Fixes #1891