Change cronjob for fetching ZA committee appearances to pombola.
Shift the committee import bits to pombola rather than za-hansard
as they need to check things in the database for name resolution.

This needs a new version of za-hansard. When we're happy that can
be merged onto master we'll need to update the requirements here
with the new commit hash.