Allow for hansard doc dates that aren't on the first line of the date
It used to be the case that hansard documents always had the date of the
document on its first line. However, newer documents don't have the date on
its first line anymore. The first few lines now contain the title and council.

In this commit we allow for the hansard date to be the first date found anywhere
in the document.